LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/basic-upgrade-cors.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<body>
<p>This test opens a window that loads an insecure image.  We should upgrade
this request and thereby avoid triggering a mixed content callback.</p>
<iframe src="https://127.0.0.1:8443/security/contentSecurityPolicy/upgrade-insecure-requests/resources/basic-upgrade-cors.https.html"></iframe>
<script>
fetch_tests_from_window(window.frames[0]);
</script>
</body>
</html>
