LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/basic-upgrade-cors.https-expected.txt

CONSOLE MESSAGE: PASS: Successfully retrieved image data.
This test opens a window that loads an insecure image. We should upgrade this request and thereby avoid triggering a mixed content callback.



PASS Verify that images have correct cross-origin behavior.
